// packages/simulator-core/src/protocol/runtime/ocpp16/types.ts
export interface Ocpp16StartTransactionInput {
  connectorId: number;
  idTag: string;
  meterStartWh: number;
  reservationId?: number;
  startedAt?: Date;
}